Refactored keep/delete moderation queue code. Now works with image or game reports. Fixed bug with moderation queue where sometimes the report would not be cleared.

+++ scripts/reportqueue_keep.php
@@ -5,6 +5,7 @@
 	// Script to keep an image that was reported for moderation
 	// Parameters: 
 	//	reportType - The type of the item in the moderation queue
+	// 	reportedID - The ID of the item that was reported
 	// 	reportID - The ID of the item in the moderation queue
 	
 	include("../include.php");
@@ -12,43 +13,59 @@
 	
 	if ($loggedin = 1 && $adminuserlevel = 'ADMINISTRATOR')
 	{
-		if (isset($reportID) and isset($reportType))
-		{
-			if ($reportedResult = mysql_query("SELECT b.filename FROM moderation_reported AS m, banners AS b WHERE m.bannerid = b.id AND m.id = $reportID LIMIT 1"))
-			{	
-				if (mysql_query("DELETE FROM moderation_reported WHERE id = $reportID"))
+		if (isset($reportID) and isset($reportType) and isset($reportedID))
+		{	
+			if (mysql_query("DELETE FROM moderation_reported WHERE id = $reportID"))
+			{
+				echo "Deleted report succesfully.\n";
+
+				// Start keep - Images
+				// ========================
+				if ($reportType == "image")
 				{
-					$reported = mysql_fetch_object($reportedResult);
-					if (isset($reported->filename))
+					// Gather info to clear from cache, but keep the image data	
+					if ($reportedResult == mysql_query("SELECT b.filename FROM banners WHERE b.id == $reportedID LIMIT 1"))
 					{
-						if(file_exists("../reportedqueue/_cache/$reported->filename"))
+						$reported = mysql_fetch_object($reportedResult);
+
+						// Find the filename
+						if (isset($reported->filename))
 						{
-							unlink("../reportedqueue/_cache/$reported->filename");
-							echo "Success";
+							// Remove the reported image from the report cache
+							if(file_exists("../reportedqueue/_cache/$reported->filename"))
+							{
+								unlink("../reportedqueue/_cache/$reported->filename");
+								echo "Successfully kept image in database.";
+							}
+							else
+							{
+								echo "Error: The image was removed from the reported list succesfully, however we were unable to locate the reported cache file for this image";
+							}
 						}
 						else
 						{
-							echo "Error: The image was removed from the reported list succesfully, however we were unable to locate the reported cache file for this image";
+							echo "Error: Unable to locate the filename of existing image from the image database";
 						}
 					}
-					else
-					{
-						echo "Error: Unable to locate the filename of existing image from the image database";
-					}
 				}
-				else
+				// ========================
+
+				// Start keep - Games
+				// ========================
+				if ($reportType == "game")
 				{
-					echo "Error: Unable to remove the reported image from the moderation database table";
+					// Nothing to do here.
+					echo "Successfully kept game in database.";
 				}
 			}
 			else
 			{
-				echo "Error: Unable to look up existing image in the images database";
+				echo "Error: Unable to remove the reported $reportType from the moderation database table. Most likely was deleted already.";
 			}
 		}
 		else
 		{
-			echo "Error: You must provide a the ID and type of the reported image/game to keep.";
+			echo "Error: You must provide a the report ID, ID to keep, and type of the reported image/game.";
 		}
 	}
 	else
